Move utils to util folder
