+++ src/sort/util/has-sort-direction.ts
@@ -1,0 +1,5 @@
+import { SortDirections } from "../sort.interface";
+
+export const hasSortDirection = (sortDirection: SortDirections): boolean => {
+    return sortDirection !== SortDirections.None;
+};
